requirement.txt
discord.py==2.6.0
ollama
Pillow
aiohttp
torch
TTS
